pyproject.toml
[tool.poetry]
name = "drf-hackathon-template"
version = "0.1.0"
description = ""
authors = ["HCodeKeeper <ivannikovmax04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
drf-spectacular = "^0.27.1"
isort = "^5.13.2"
bandit = "^1.7.7"
black = "^24.2.0"
flake8 = "^7.0.0"
pre-commit = "^3.6.2"
pytest = "^8.0.2"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"
uvicorn = "^0.27.1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
